Early draft of mockup
Server application for grocery shopping where it keeps track of what’s in the fridge, makes a grocery list based on what meal to make and what’s in the fridge. Uses REST APIs to connect a frontend application or mobile application.
Goal for tomorrow:
A more detailed description posted on the README.
A fully completed mockup.